+++ packages/patternfly-4/react-core/src/components/Tabs/Tab.tsx
@@ -18,5 +18,4 @@ export interface TabProps extends Omit<React.HTMLProps<HTMLAnchorElement | HTMLB
   tabContentRef?: React.RefObject<any>;
 }
 
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
 export const Tab: React.FunctionComponent<TabProps> = ({ className = '' }: TabProps) => null;