+++ packages/patternfly-4/react-integration/demo-app-ts/src/components/demos/SwitchDemo/SwitchDemo.tsx
@@ -1,26 +1,55 @@
 import React from 'react';
-import { Switch } from '@patternfly/react-core';
+import { Switch, Button } from '@patternfly/react-core';
+import { Table, TableBody, TableHeader } from '@patternfly/react-table';
 
 interface SwitchState {
-  isChecked: boolean;
+  simple: {
+    isChecked: boolean;
+  };
+  table: {
+    data: { id: string; isChecked: boolean }[];
+    all: boolean;
+  };
 }
+
 export class SwitchDemo extends React.Component<{}, SwitchState> {
   state = {
-    isChecked: true
+    simple: {
+      isChecked: true
+    },
+    table: {
+      data: [{ id: 'id1', isChecked: false }, { id: 'id2', isChecked: false }, { id: 'id3', isChecked: false }],
+      all: true
+    }
   };
 
-  handleChange = isChecked => {
-    this.setState({ isChecked });
+  handleChangeSimple = isChecked => {
+    this.setState({ simple: { isChecked } });
   };
 
-  render() {
-    const { isChecked } = this.state;
+  handleChangeTable = id => isChecked => {
+    this.setState({
+      table: {
+        ...this.state.table,
+        data: this.state.table.data.map(d => (d.id === id ? { ...d, isChecked } : d))
+      }
+    });
+  };
+
+  showAllTable = () => {
+    this.setState({ table: { ...this.state.table, all: !this.state.table.all } });
+  };
+
+  renderSimple() {
+    const {
+      simple: { isChecked }
+    } = this.state;
     return (
       <React.Fragment>
         <Switch
           id="simple-switch"
           label={isChecked ? 'Message when on' : 'Message when off'}
-          onChange={this.handleChange}
+          onChange={this.handleChangeSimple}
           aria-label="Switch"
           isChecked={isChecked}
         />
@@ -32,6 +61,46 @@ export class SwitchDemo extends React.Component<{}, SwitchState> {
           isChecked={false}
           isDisabled
         />
+      </React.Fragment>
+    );
+  }
+
+  renderTable() {
+    const cols = [{ title: 'hello' }];
+    const {
+      table: { data, all }
+    } = this.state;
+    let rr = data;
+    if (!all) {
+      // eslint-disable-next-line @typescript-eslint/no-unused-vars
+      const [first, ...rest] = data;
+      rr = rest;
+    }
+    const rows = rr.map(r => ({
+      cells: [
+        {
+          title: <Switch id={r.id} isChecked={r.isChecked} onChange={this.handleChangeTable(r.id)} />
+        }
+      ]
+    }));
+    return (
+      <div>
+        <Table aria-label="simple table" cells={cols} rows={rows}>
+          <TableHeader />
+          <TableBody />
+        </Table>
+        <Button id="showAll" onClick={this.showAllTable}>
+          Show all
+        </Button>
+      </div>
+    );
+  }
+
+  render() {
+    return (
+      <React.Fragment>
+        {this.renderSimple()}
+        {this.renderTable()}
       </React.Fragment>
     );
   }