+++ packages/patternfly-4/react-catalog-view-extension/CHANGELOG.md
@@ -3,6 +3,81 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.20](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@1.4.19...@patternfly/react-catalog-view-extension@1.4.20) (2020-03-02)
+
+
+### Bug Fixes
+
+* **CatalogTile:** Removed truncation and maxLength props ([#3830](https://github.com/patternfly/patternfly-react/issues/3830)) ([e0402d0](https://github.com/patternfly/patternfly-react/commit/e0402d02a2af703e7520f521da91ca39ef1a8a64)), closes [#3471](https://github.com/patternfly/patternfly-react/issues/3471)
+
+
+
+
+
+## [1.4.19](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@1.4.18...@patternfly/react-catalog-view-extension@1.4.19) (2020-03-02)
+
+**Note:** Version bump only for package @patternfly/react-catalog-view-extension
+
+
+
+
+
+## [1.4.18](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@1.4.17...@patternfly/react-catalog-view-extension@1.4.18) (2020-02-28)
+
+**Note:** Version bump only for package @patternfly/react-catalog-view-extension
+
+
+
+
+
+## [1.4.17](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@1.4.16...@patternfly/react-catalog-view-extension@1.4.17) (2020-02-27)
+
+**Note:** Version bump only for package @patternfly/react-catalog-view-extension
+
+
+
+
+
+## [1.4.16](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@1.4.15...@patternfly/react-catalog-view-extension@1.4.16) (2020-02-26)
+
+**Note:** Version bump only for package @patternfly/react-catalog-view-extension
+
+
+
+
+
+## [1.4.15](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@1.4.14...@patternfly/react-catalog-view-extension@1.4.15) (2020-02-26)
+
+**Note:** Version bump only for package @patternfly/react-catalog-view-extension
+
+
+
+
+
+## [1.4.14](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@1.4.13...@patternfly/react-catalog-view-extension@1.4.14) (2020-02-25)
+
+**Note:** Version bump only for package @patternfly/react-catalog-view-extension
+
+
+
+
+
+## [1.4.13](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@1.4.12...@patternfly/react-catalog-view-extension@1.4.13) (2020-02-25)
+
+**Note:** Version bump only for package @patternfly/react-catalog-view-extension
+
+
+
+
+
+## [1.4.12](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@1.4.11...@patternfly/react-catalog-view-extension@1.4.12) (2020-02-25)
+
+**Note:** Version bump only for package @patternfly/react-catalog-view-extension
+
+
+
+
+
 ## [1.4.11](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-catalog-view-extension@1.4.10...@patternfly/react-catalog-view-extension@1.4.11) (2020-02-18)
 
 **Note:** Version bump only for package @patternfly/react-catalog-view-extension