+++ packages/patternfly-4/react-integration/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.25.1](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-integration@3.25.0...@patternfly/react-integration@3.25.1) (2020-02-26)
+
+
+### Bug Fixes
+
+* **switch:** avoid switch id override by props ([#3706](https://github.com/patternfly/patternfly-react/issues/3706)) ([0952774](https://github.com/patternfly/patternfly-react/commit/0952774d3dbcf6a7d66d4463277fab181bfd6c28))
+
+
+
+
+
+# [3.25.0](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-integration@3.24.3...@patternfly/react-integration@3.25.0) (2020-02-25)
+
+
+### Features
+
+* **card:** add selectable and selected variation ([#3587](https://github.com/patternfly/patternfly-react/issues/3587)) ([166aa09](https://github.com/patternfly/patternfly-react/commit/166aa097ae08f5da216c9f59a1094e032da49054))
+
+
+
+
+
 ## [3.24.3](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-integration@3.24.2...@patternfly/react-integration@3.24.3) (2020-02-17)