+++ packages/patternfly-4/react-styles/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.7.5](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-styles@3.7.4...@patternfly/react-styles@3.7.5) (2020-02-27)
+
+**Note:** Version bump only for package @patternfly/react-styles
+
+
+
+
+
 ## [3.7.4](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-styles@3.7.3...@patternfly/react-styles@3.7.4) (2020-02-18)
 
 **Note:** Version bump only for package @patternfly/react-styles