+++ packages/patternfly-4/react-topology/CHANGELOG.md
@@ -3,6 +3,78 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.14.20](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@2.14.19...@patternfly/react-topology@2.14.20) (2020-03-02)
+
+**Note:** Version bump only for package @patternfly/react-topology
+
+
+
+
+
+## [2.14.19](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@2.14.18...@patternfly/react-topology@2.14.19) (2020-03-02)
+
+**Note:** Version bump only for package @patternfly/react-topology
+
+
+
+
+
+## [2.14.18](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@2.14.17...@patternfly/react-topology@2.14.18) (2020-02-28)
+
+**Note:** Version bump only for package @patternfly/react-topology
+
+
+
+
+
+## [2.14.17](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@2.14.16...@patternfly/react-topology@2.14.17) (2020-02-27)
+
+**Note:** Version bump only for package @patternfly/react-topology
+
+
+
+
+
+## [2.14.16](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@2.14.15...@patternfly/react-topology@2.14.16) (2020-02-26)
+
+**Note:** Version bump only for package @patternfly/react-topology
+
+
+
+
+
+## [2.14.15](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@2.14.14...@patternfly/react-topology@2.14.15) (2020-02-26)
+
+**Note:** Version bump only for package @patternfly/react-topology
+
+
+
+
+
+## [2.14.14](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@2.14.13...@patternfly/react-topology@2.14.14) (2020-02-25)
+
+**Note:** Version bump only for package @patternfly/react-topology
+
+
+
+
+
+## [2.14.13](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@2.14.12...@patternfly/react-topology@2.14.13) (2020-02-25)
+
+**Note:** Version bump only for package @patternfly/react-topology
+
+
+
+
+
+## [2.14.12](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@2.14.11...@patternfly/react-topology@2.14.12) (2020-02-25)
+
+**Note:** Version bump only for package @patternfly/react-topology
+
+
+
+
+
 ## [2.14.11](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@2.14.10...@patternfly/react-topology@2.14.11) (2020-02-18)
 
 **Note:** Version bump only for package @patternfly/react-topology